Refactored and renamed 'base/baseAnnotationToolHelpers.js' to 'util/findAndMoveHelpers.js' and 'store/getToolsWithMoveableHandles.js', as the functionality is used by more than just the baseAnnotationTool.
