+++ src/util/findAndMoveHelpers.js
@@ -1,29 +1,7 @@
 import { mutations, state } from '../store/index.js';
-import { getToolState } from '../stateManagement/toolState.js';
 import getHandleNearImagePoint from '../manipulators/getHandleNearImagePoint.js';
 import moveAllHandles from '../manipulators/moveAllHandles.js';
 import moveHandle from '../manipulators/moveHandle.js';
-
-const getToolsWithMovableHandles = function (element, tools, coords) {
-  return tools.filter((tool) => {
-    const toolState = getToolState(element, tool.name);
-
-    for (let i = 0; i < toolState.data.length; i++) {
-      if (
-        getHandleNearImagePoint(
-          element,
-          toolState.data[i].handles,
-          coords,
-          state.clickProximity
-        ) !== undefined
-      ) {
-        return true;
-      }
-    }
-
-    return false;
-  });
-};
 
 const moveHandleNearImagePoint = function (evt, handle, data, toolName) {
   // Todo: We've grabbed a handle, stop listening/ignore for MOUSE_MOVE
@@ -114,7 +92,6 @@ const findAnnotationNearClick = function (
 };
 
 export {
-  getToolsWithMovableHandles,
   moveHandleNearImagePoint,
   findHandleDataNearImagePoint,
   moveAnnotationNearClick,
